Initial import.
This code was originally part of https://github.com/nstockton/mapperproxy-mume.
I decided to split this code off into its own repository so I can reuse it for future projects.
To see the commit history prior to this commit, look at the mapper/protocols directory of mapperproxy-mume, starting with this commit:
https://github.com/nstockton/mapperproxy-mume/commit/a02ee21cbded56831c22c5d0a838039f7c2e3b2a